requirements.txt
certifi==2019.6.16
Click==7.0
Flask==1.1.1
flask-marshmallow==0.10.1
Flask-SQLAlchemy==2.4.0
gunicorn==19.9.0
itsdangerous==1.1.0
Jinja2==2.10.1
MarkupSafe==1.1.1
marshmallow==2.19.5
marshmallow-sqlalchemy==0.17.0
psycopg2==2.8.3
python-dotenv==0.10.3
six==1.12.0
SQLAlchemy==1.3.6
Werkzeug==0.15.5
